Add more tests and fix lint
