+++ android/beagle/src/main/java/br/com/zup/beagle/android/context/tokenizer/function/builtin/array/InsertFunction.kt
@@ -23,6 +23,7 @@ import org.json.JSONArray
 internal class InsertFunction : Function {
     override fun functionName(): String = "insert"
 
+    @Suppress("ReturnCount")
     override fun execute(vararg params: Any?): Any {
         val array = params[0]
         val element = params[1] as Any
