+++ android/beagle/src/main/java/br/com/zup/beagle/android/context/tokenizer/function/builtin/array/RemoveFunction.kt
@@ -22,6 +22,7 @@ import org.json.JSONArray
 internal class RemoveFunction : Function {
     override fun functionName(): String = "remove"
 
+    @Suppress("ReturnCount")
     override fun execute(vararg params: Any?): Any {
         val array = params[0]
         val element = params[1] as Any
